# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KCardSwap API",
    "description": "Backend API for KCardSwap - Card Exchange Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chats": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get chat rooms",
        "description": "Get all chat rooms for the current user",
        "operationId": "get_chat_rooms_api_v1_chats_get",
        "responses": {
          "200": {
            "description": "Chat rooms retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatRoomResponse"
                  },
                  "type": "array",
                  "title": "Response Get Chat Rooms Api V1 Chats Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chats/{room_id}/messages": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get messages",
        "description": "Get messages from a chat room (with pagination support via after_message_id)",
        "operationId": "get_messages_api_v1_chats__room_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          },
          {
            "name": "after_message_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Get messages after this message ID",
              "title": "After Message Id"
            },
            "description": "Get messages after this message ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of messages",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of messages"
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "403": {
            "description": "Forbidden (not a participant of this room)"
          },
          "404": {
            "description": "Chat room not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Send message",
        "description": "Send a message in a chat room (triggers FCM push notification)",
        "operationId": "send_message_api_v1_chats__room_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (validation failed)"
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "403": {
            "description": "Forbidden (not authorized to send message)"
          },
          "404": {
            "description": "Chat room not found"
          },
          "422": {
            "description": "Unprocessable entity (blocked or not friends)"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/chats/{room_id}/messages/{message_id}/read": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Mark message as read",
        "description": "Mark a message as read by the current user",
        "operationId": "mark_message_read_api_v1_chats__room_id__messages__message_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Message marked as read"
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "403": {
            "description": "Forbidden (not authorized)"
          },
          "404": {
            "description": "Message not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRoomParticipantResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "User ID"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname",
            "description": "User's nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "description": "User's avatar URL"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ChatRoomParticipantResponse",
        "description": "Response schema for chat room participant"
      },
      "ChatRoomResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Chat room ID"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ChatRoomParticipantResponse"
            },
            "type": "array",
            "title": "Participants",
            "description": "Room participants"
          },
          "last_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last message in the room"
          },
          "unread_count": {
            "type": "integer",
            "title": "Unread Count",
            "description": "Number of unread messages for current user",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "participants",
          "created_at"
        ],
        "title": "ChatRoomResponse",
        "description": "Response schema for chat room details",
        "example": {
          "created_at": "2024-01-01T00:00:00Z",
          "id": "987e6543-e21b-12d3-a456-426614174000",
          "last_message": {
            "content": "Hello!",
            "created_at": "2024-01-01T00:00:00Z",
            "id": "123e4567-e89b-12d3-a456-426614174000",
            "room_id": "987e6543-e21b-12d3-a456-426614174000",
            "sender_id": "456e7890-e12b-34d5-a678-901234567000",
            "status": "read"
          },
          "participants": [
            {
              "avatar_url": "https://example.com/avatar.jpg",
              "nickname": "John Doe",
              "user_id": "456e7890-e12b-34d5-a678-901234567000"
            },
            {
              "avatar_url": "https://example.com/avatar2.jpg",
              "nickname": "Jane Smith",
              "user_id": "789e0123-e45b-67d8-a901-234567890000"
            }
          ],
          "unread_count": 0
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Message ID"
          },
          "room_id": {
            "type": "string",
            "format": "uuid",
            "title": "Room Id",
            "description": "Chat room ID"
          },
          "sender_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sender Id",
            "description": "Sender user ID"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Message content"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Message status",
            "examples": [
              "sent",
              "delivered",
              "read"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "room_id",
          "sender_id",
          "content",
          "status",
          "created_at"
        ],
        "title": "MessageResponse",
        "description": "Response schema for a single message",
        "example": {
          "content": "Hello! I'm interested in trading cards.",
          "created_at": "2024-01-01T00:00:00Z",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "room_id": "987e6543-e21b-12d3-a456-426614174000",
          "sender_id": "456e7890-e12b-34d5-a678-901234567000",
          "status": "sent"
        }
      },
      "MessagesListResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "type": "array",
            "title": "Messages",
            "description": "List of messages"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of messages"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether there are more messages to fetch"
          }
        },
        "type": "object",
        "required": [
          "messages",
          "total",
          "has_more"
        ],
        "title": "MessagesListResponse",
        "description": "Response schema for list of messages",
        "example": {
          "has_more": false,
          "messages": [
            {
              "content": "Hello!",
              "created_at": "2024-01-01T00:00:00Z",
              "id": "123e4567-e89b-12d3-a456-426614174000",
              "room_id": "987e6543-e21b-12d3-a456-426614174000",
              "sender_id": "456e7890-e12b-34d5-a678-901234567000",
              "status": "read"
            }
          ],
          "total": 1
        }
      },
...
      "SendMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "description": "Message content",
            "examples": [
              "Hello! I'm interested in trading cards."
            ]
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "SendMessageRequest",
        "description": "Request schema for sending a message",
        "example": {
          "content": "Hello! I'm interested in trading cards."
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}